+++ wp-content/plugins/analytics-for-wp/options.php
@@ -1,14 +1,28 @@
-<div class="wrap">
+<div class="wrap" id="gaw">
 <h2>Analytics for WP Settings</h2>
 
 <?php
+
+    $options = get_option('gaw_analytics_id');
+    $result =isAnalytics($options); 
+    
    
 if( isset($_GET['settings-updated']) && $_GET['settings-updated'] == 'true'):
-   echo    '<div id="setting-error-settings_updated" class="updated settings-error"> 
-<p><strong>Settings saved.</strong></p></div>';
+    if($result){
+        echo '<div class="updated notice notice-success is-dismissible"> 
+                <p><strong>Settings saved.</strong></p>
+            </div>';        
+    }else{
+        echo '<div class="notice notice-error is-dismissible"> 
+                <p><strong>Invalid GA code. Please enter valid code.</strong></p>
+            </div>';
+        update_option('gaw_analytics_id', '');
+    }
+    
 endif;
 ?>
 
+    
    <form method="post" action="options.php">
        <?php settings_fields( 'gaw-settings-group' ); ?>
        <?php do_settings_sections( 'gaw-settings-group' ); ?>
@@ -16,14 +30,14 @@ endif;
           <tr valign="top">
             <th scope="row">Tracking Code</th>
             <td>
-               <input type="text" style="width:50%" name="gaw_analytics_id" value="<?php echo get_option('gaw_analytics_id', 'UA-XXXXXXXX-1'); ?>" />
+               <input type="text" style="width:50%" name="gaw_analytics_id" value="<?php echo get_option('gaw_analytics_id', ''); ?>" placeholder="UA-XXXXXXXX-1" />
                <br/><small>If you  don't have tracking code <a href="https://analytics.google.com/" target="_blank">click here.</a></small>
             </td>
           </tr>
           <tr valign="top">
              <th scope="row">Disable Tracking</th>
              <td>
-               <select name="gaw_disable_track" />
+               <select name="gaw_disable_track">
                    <option value="No" <?php if( get_option('gaw_disable_track') == "No" ): echo 'selected'; endif;?> >No</option>
                    <option value="Yes" <?php if( get_option('gaw_disable_track') == "Yes" ): echo 'selected'; endif;?> >Yes</option>
                </select>
@@ -35,4 +49,7 @@ endif;
 
        <p>For any help <a href="https://amanvermaa.wordpress.com/2017/02/25/how-to-add-google-analytics-tracking-code-in-wordpress-website/?ref=wp-analytics" target="_blank">click here</a></p>
    </form>
+    <div>
+        <h3><a href="https://wordpress.org/support/plugin/analytics-for-wp/reviews/?filter=5#new-post">Please review us</a> if you like the plugin.</h3>
+    </div>
 </div>