+++ wp-content/plugins/themeisle-companion/core/app/class-orbit-fox-global-settings.php
@@ -62,7 +62,6 @@ class Orbit_Fox_Global_Settings {
 				'obfx_modules',
 				array(
 					'social-sharing',
-					'image-cdn',
 					'gutenberg-blocks',
 					'uptime-monitor',
 					'google-analytics',
@@ -73,7 +72,7 @@ class Orbit_Fox_Global_Settings {
 					'mystock-import',
 					'policy-notice',
 					'beaver-widgets',
-					'safe-updates',
+					'image-cdn',
 				)
 			);
 		}// End if().