+++ wp-content/plugins/themeisle-companion/obfx_modules/mystock-import/init.php
@@ -23,14 +23,14 @@ class Mystock_Import_OBFX_Module extends Orbit_Fox_Module_Abstract {
 	const API_KEY = '97d007cf8f44203a2e578841a2c0f9ac';
 
 	/**
+	 * Flickr user id.
+	 */
+	const USER_ID = '136375272@N05';
+
+	/**
 	 * The number of images to fetch. Only the first page will be fetched.
 	 */
 	const MAX_IMAGES = 40;
-
-	/**
-	 * The username of the flickr account.
-	 */
-	const USER_NAME = 'themeisle';
 
 	/**
 	 * The cache time.
@@ -55,9 +55,9 @@ class Mystock_Import_OBFX_Module extends Orbit_Fox_Module_Abstract {
 	/**
 	 * Determine if module should be loaded.
 	 *
+	 * @return bool
 	 * @since   1.0.0
 	 * @access  public
-	 * @return bool
 	 */
 	public function enable_module() {
 		return true;
@@ -92,6 +92,7 @@ class Mystock_Import_OBFX_Module extends Orbit_Fox_Module_Abstract {
 	 */
 	public function get_tab_content() {
 		$urls = $this->get_images();
+		$page = 1;
 		require $this->get_dir() . "/inc/photos.php";
 		wp_die();
 	}
@@ -108,17 +109,13 @@ class Mystock_Import_OBFX_Module extends Orbit_Fox_Module_Abstract {
 		if ( ! $photos ) {
 			require_once $this->get_dir() . '/vendor/phpflickr/phpflickr.php';
 			$api    = new phpFlickr( self::API_KEY );
-			$user   = $api->people_findByUsername( self::USER_NAME );
-			$photos = array();
-			if ( $user && isset( $user['nsid'] ) ) {
-				$photos = $api->people_getPublicPhotos( $user['nsid'], null, 'url_sq, url_t, url_s, url_q, url_m, url_n, url_z, url_c, url_l, url_o', self::MAX_IMAGES, $page );
-				if ( ! empty( $photos ) ) {
-					$pages = get_transient( $this->slug . 'photos_' . self::MAX_IMAGES . '_pages' );
-					if ( false === $pages ) {
-						set_transient( $this->slug . 'photos_' . self::MAX_IMAGES . '_pages', $photos['photos']['pages'], self::CACHE_DAYS * DAY_IN_SECONDS );
-					}
-					$photos = $photos['photos']['photo'];
+			$photos = $api->people_getPublicPhotos( self::USER_ID, null, 'url_sq, url_t, url_s, url_q, url_m, url_n, url_z, url_c, url_l, url_o', self::MAX_IMAGES, $page );
+			if ( ! empty( $photos ) ) {
+				$pages = get_transient( $this->slug . 'photos_' . self::MAX_IMAGES . '_pages' );
+				if ( false === $pages ) {
+					set_transient( $this->slug . 'photos_' . self::MAX_IMAGES . '_pages', $photos['photos']['pages'], self::CACHE_DAYS * DAY_IN_SECONDS );
 				}
+				$photos = $photos['photos']['photo'];
 			}
 			set_transient( $this->slug . 'photos_' . self::MAX_IMAGES . '_' . $page, $photos, self::CACHE_DAYS * DAY_IN_SECONDS );
 		}
@@ -173,23 +170,15 @@ class Mystock_Import_OBFX_Module extends Orbit_Fox_Module_Abstract {
 		}
 
 		//Update last page that was loaded
-		$req_page = (int) $_POST['page'] + 1;
+		$page = (int) $_POST['page'] + 1;
 
 		//Request new page
-		$response    = '';
-		$new_request = $this->get_images( $req_page );
-		if ( ! empty( $new_request ) ) {
-			foreach ( $new_request as $photo ) {
-				$response .= '<li class="obfx-image" data-page="' . esc_attr( $req_page ) . '" data-pid="' . esc_attr( $photo['id'] ) . '">';
-				$response .= '<div class="obfx-preview"><div class="thumbnail"><div class="centered">';
-				$response .= '<img src="' . esc_url( $photo['url_m'] ) . '">';
-				$response .= '</div></div></div>';
-				$response .= '<button type="button" class="check obfx-image-check" tabindex="0"><span class="media-modal-icon"></span><span class="screen-reader-text">' . esc_html__( 'Deselect', 'themeisle-companion' ) . '</span></button>';
-				$response .= '</li>';
+		$urls = $this->get_images( $page );
+		if ( ! empty( $urls ) ) {
+			foreach ( $urls as $photo ) {
+				include $this->get_dir() . '/inc/photo.php';
 			}
 		}
-
-		echo $response;
 		wp_die();
 	}
 
@@ -197,9 +186,9 @@ class Mystock_Import_OBFX_Module extends Orbit_Fox_Module_Abstract {
 	 * Method that returns an array of scripts and styles to be loaded
 	 * for the front end part.
 	 *
+	 * @return array
 	 * @since   1.0.0
 	 * @access  public
-	 * @return array
 	 */
 	public function public_enqueue() {
 		return array();
@@ -209,13 +198,12 @@ class Mystock_Import_OBFX_Module extends Orbit_Fox_Module_Abstract {
 	 * Method that returns an array of scripts and styles to be loaded
 	 * for the admin part.
 	 *
+	 * @return array
 	 * @since   1.0.0
 	 * @access  public
-	 * @return array
 	 */
 	public function admin_enqueue() {
 		$current_screen = get_current_screen();
-
 		if ( ! isset( $current_screen->id ) ) {
 			return array();
 		}
@@ -256,9 +244,9 @@ class Mystock_Import_OBFX_Module extends Orbit_Fox_Module_Abstract {
 	/**
 	 * Method to define the options fields for the module
 	 *
+	 * @return array
 	 * @since   1.0.0
 	 * @access  public
-	 * @return array
 	 */
 	public function options() {
 		return array();